+++ examples/awaaz-chatbot/env.example
@@ -1,0 +1,3 @@
+OPENAI_API_KEY=
+DEEPGRAM_API_KEY=
+ELEVENLABS_API_KEY=
