+++ examples/awaaz-chatbot/templates/streams.xml.template
