--- examples/awaazai-patient-intake/env.example
@@ -1,3 +1,0 @@
-OPENAI_API_KEY=sk-PL...
-CARTESIA_API_KEY=your_cartesia_api_key_here
-DEEPGRAM_API_KEY=63...
